NotificationListenerViewController super class
