+++ mokojin/JoinQueueViewController.swift
@@ -9,60 +9,41 @@
 import Foundation
 import UIKit
 
-class JoinQueueViewController : UIViewController, UITableViewDataSource, UITableViewDelegate {
+class JoinQueueViewController : NotificationListenerViewController, UITableViewDataSource, UITableViewDelegate {
     
     @IBOutlet var tableView: UITableView!
     var people:People = []
-    let notificationManager = NotificationManager()
-    
-    override init() {
-        super.init()
-        self.listenOnEvents()
-    }
     
     required init(coder: NSCoder) {
         super.init(coder: coder)
-        self.listenOnEvents()
     }
-    
-    private func listenOnEvents(){
-        notificationManager.registerObserver(PeopleStore.sharedInstance.NOTIFICATION_NAME, block: { notification in
-            self.updateData()
-        })
-    }
-    
-    deinit {
-        notificationManager.deregisterAll()
-    }
-    
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        tableView.delegate = self
-        tableView.dataSource = self
-        updateData()
+        self.tableView.delegate = self
+        self.tableView.dataSource = self
     }
     
+    override func getNotificationName() -> String {
+        return PeopleStore.sharedInstance.NOTIFICATION_NAME
+    }
     
-    private func updateData(){
+    override func updateData(){
         self.people = PeopleStore.sharedInstance.people
-        tableView.reloadData()
+        self.tableView.reloadData()
     }
     
-
     func numberOfSectionsInTableView(tableView: UITableView) -> Int {
         return 1
     }
     
     func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return people.count
+        return self.people.count
     }
     
     func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCellWithIdentifier("PlayerCellID", forIndexPath: indexPath) as UITableViewCell
-        
         cell.textLabel?.text = people[indexPath.row].name
-        
         return cell
     }
